files/SANP_DOCS/requirements.txt
mkdocs-material==9.5.49
mkdocs-git-revision-date-localized-plugin==1.3.0
mkdocs-glightbox==0.4.0

# mkdocs-git-authors-plugin==0.8.0
# mkdocs-git-committers-plugin-2==2.3.0